rename config to common
