+++ sktime/split/base/_base_splitter.py
@@ -9,7 +9,7 @@ import pandas as pd
 from sktime.base import BaseObject
 from sktime.datatypes import check_is_scitype, convert
 from sktime.forecasting.base import ForecastingHorizon
-from sktime.split.base._config import (
+from sktime.split.base._common import (
     ACCEPTED_Y_TYPES,
     DEFAULT_FH,
     DEFAULT_WINDOW_LENGTH,
