+++ sktime/split/base/_base_windowsplitter.py
@@ -9,7 +9,7 @@ import pandas as pd
 from sktime.datatypes._utilities import get_index_for_series
 from sktime.forecasting.base import ForecastingHorizon
 from sktime.split.base import BaseSplitter
-from sktime.split.base._config import (
+from sktime.split.base._common import (
     ACCEPTED_Y_TYPES,
     FORECASTING_HORIZON_TYPES,
     SPLIT_ARRAY_TYPE,
