+++ sktime/split/expandingwindow.py
@@ -7,7 +7,7 @@ __all__ = [
 ]
 
 from sktime.split.base import BaseWindowSplitter
-from sktime.split.base._config import (
+from sktime.split.base._common import (
     DEFAULT_FH,
     DEFAULT_STEP_LENGTH,
     DEFAULT_WINDOW_LENGTH,
