+++ sktime/split/testplustrain.py
@@ -10,7 +10,7 @@ import numpy as np
 import pandas as pd
 
 from sktime.split.base import BaseSplitter
-from sktime.split.base._config import ACCEPTED_Y_TYPES, SPLIT_GENERATOR_TYPE
+from sktime.split.base._common import ACCEPTED_Y_TYPES, SPLIT_GENERATOR_TYPE
 
 
 class TestPlusTrainSplitter(BaseSplitter):
