+++ sktime/split/tests/test_expandingwindow.py
@@ -12,7 +12,7 @@ from sktime.forecasting.tests._config import (
     TEST_YS,
 )
 from sktime.split import ExpandingWindowSplitter
-from sktime.split.base._config import _inputs_are_supported
+from sktime.split.base._common import _inputs_are_supported
 from sktime.split.tests.test_split import _check_cv, _get_n_incomplete_windows
 from sktime.utils.datetime import _coerce_duration_to_int
 from sktime.utils.validation.forecasting import check_fh
